refactor(runtime): inline lifecycle into message adapter
